# OrchestratorService/orchestratorservice-swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Orchestrator Service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Projects": {
      "post": {
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Team/projects/{projectId}": {
      "get": {
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "required": [
          "orchestratorName",
          "projectName",
          "repositoryAddress",
          "repositoryName"
        ],
        "type": "object",
        "properties": {
          "projectName": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "repositoryName": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "repositoryAddress": {
            "minLength": 1,
            "type": "string",
            "format": "uri"
          },
          "orchestratorName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}